Fix NSApplication.SharedApplication
We currently populate the address space in the NSObject.cctor, but
Class.GetHandle ("NSApplication") would fail since its not paged in
yet.  So we're going to force the NSObject.cctor to run in the
NSApplication.cctor, and repopulate the class_ptr in the .cctor
so that we have a valid class connection.  I'm still not entirely
happy with this approach, and would like to bust each set of dependent
libraries into their own module and have the native loader do the fixup
in a module initializer, but that would require some compiler love.