Further byref fixes.
When we are marshalling byref argument we were not checking the element type
properly, which could lead to a marshalling directive exception.  To fix this
we need to store the expected types of the method as well as the marshal types
so we can properly emit our conversion routines in the trampolines for the
byref case.