Merging the additions from the version of MonoMac that TouchDraw uses into the master.
Most of the additions (missing mappings) I had made are now in the
current version of MonoMac.  These are the only two additions/changes
not in the master:

NSDocument.FileUrl can be set to null (and TouchDraw needs to do this),
so added a NullAllowed for it.

Changed the signature of NSDocument.SetDisplayName(string) to
SetDisplayName(NSString).  I did this because if that method was
overridden (which TouchDraw does) and a null value was passed into that
method from the Objective-C side, MonoMac will crash.