NSAlert: more BeginSheet overrides + RunSheetModal
Add BeginSheet(NSWindow, Action<int>), which is actually useful.
The returned integer is the button response code (the same
as NSAlert.RunModal). Previously there was only
BeginSheet(NSWindow, NSAction), which means you could never
know what the button response was, only when the sheet was
dismissed (somehow).

Additionally a convenience method, RunSheetModal is added.

The correct selector for NSAlert.BeginSheet is
alertDidEnd:returnCode:contextInfo:, unlike other BeginSheet
methods where apply is the appropriate selector.

Also make NSAlert.Window an NSPanel and allow window to be
null on the generated raw BeginSheet method. Docs say that if
the window a sheet should attach to is null, the alert will
be displayed as a dialog instead.