Add overlooked Foundation files from maccore to the build.
It appears that the following files from maccore were not included in the build. NSUbiquitousKeyValueStore in particular is needed in order to gain access to setters for iCloud key/value storage.

NSAttributedString
NSLocale
NSMutableSet
NSUbiquitousKeyValueStore
NSUuid

The NSDistributedNotificationCenter.cs file was also not included, but it does not appear to contain any needed methods and so was not added back.