[#12597] NativeMethodBuilder: don't call NSString.ctor
When a managed method is invoked by the native runtime
that returns a System.String (to be marshalled as an
NSString), if the string value is null, a crash will occur
(ArgumentNullException).

Instead of instantiating NSString, call the explicit operator
which will return null if the source System.String is null,
or NSString otherwise.